Restore signed semantics for fields the format defines as s8
Nds4j's MemBufReader.readByte() now returns an unsigned value in 0..255,
matching Buffer.readByte() and readUInt8(). Five fields here are defined by
the Gen IV formats as signed 8-bit and were relying on the old signed
behaviour, so they began parsing as large positive numbers instead:

  MoveData.priority              -7 (Trick Room, Roar) read back as 249
  ItemData.evYields              EV-reducing berries yield negative EVs
  ItemData.friendshipChangeAmounts  bitter berries lower friendship
  PokemonSpriteData.globalFrontYOffset / shadowXOffset  signed offsets

The bytes were never at risk: every write path narrows back through a byte,
so 249 and -7 both serialize to 0xF9 and the file round-trips either way.
What broke is the value. PokEditor declares the priority column's range as
{-128, 127}, so the sheet would have displayed 249 and then rejected it as
out of range on the next edit.

Each read now converts explicitly at the point of the read. PokemonSpriteData
.unknownByte is left unsigned: it is an opaque passthrough written back with
a cast, so it has no signed interpretation to preserve.

Checked and deliberately unchanged:
- ItemData hp/ppRecoveryAmount are genuinely unsigned (0xFF is the "restore
  all" sentinel) and already mask.
- LearnsetData reads its packed halfword as a signed short, but compares the
  terminator in the short domain and masks both extracted fields, so sign
  extension cannot reach either value.
- EvolutionData's halfwords never exceed 0x7FFF in practice.

Adds SignedFieldTest, which asserts both halves of the property: parsed
values carry the right sign across the whole 0..255 domain, and the bytes
still round-trip. It needs no ROM. Reverting the priority cast fails two of
its tests while the round-trip test keeps passing, which is the distinction
being drawn here.

+++ src/main/java/io/github/turtleisaac/pokeditor/formats/items/ItemData.java
@@ -150,7 +150,7 @@ public class ItemData implements GenericFileData
         evYields = new int[NUM_EV_YIELDS];
         for (int i = 0; i < NUM_EV_YIELDS; i++)
         {
-            evYields[i] = reader.readByte(); // s8
+            evYields[i] = (byte) reader.readByte(); // s8
         }
 
         hpRecoveryAmount = reader.readByte() & 0xFF;
@@ -159,7 +159,7 @@ public class ItemData implements GenericFileData
         friendshipChangeAmounts = new int[NUM_FRIENDSHIP_CHANGE_FIELDS];
         for (int i = 0; i < NUM_FRIENDSHIP_CHANGE_FIELDS; i++)
         {
-            friendshipChangeAmounts[i] = reader.readByte();
+            friendshipChangeAmounts[i] = (byte) reader.readByte(); // s8: bitter berries lower friendship
         }
 
         // getBuffer() returns everything between the read and write positions - i.e. the unread remainder
